requirements.txt
openai>=2.0.0
langchain>=1.0.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8.0
pypdf>=4.0.0
streamlit>=1.35.0
pandas>=2.0.0
plotly>=5.20.0
python-dotenv>=1.0.0
pytest>=8.0.0
